+++ inst/tinytest/test_pkg_earth.R
@@ -1,6 +1,6 @@
 # Exits
 if (!requireNamespace("earth", quietly = TRUE)) {
-  exit_file("Package earth missing")
+  exit_file("Package 'earth' missing")
 }
 
 # Load required packages
@@ -9,7 +9,7 @@ suppressMessages({
 })
 
 # Generate Friedman benchmark data
-friedman1 <- gen_friedman(seed = 101, n_bins = 2)
+friedman1 <- gen_friedman(seed = 101)
 
 # Fit model(s)
 fit <- earth::earth(y ~ ., degree = 2, data = friedman1)
@@ -44,22 +44,4 @@ expect_identical(
 expect_identical(
   current = vip:::get_feature_names.earth(fit),
   target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit, method = "model", include_type = TRUE)
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis_nsubsets, include_type = TRUE),
-  vip(vis_rss, include_type = TRUE),
-  vip(vis_gcv, include_type = TRUE),
-  p,
-  nrow = 2
 )