+++ inst/tinytest/test_pkg_glmnet.R
@@ -1,6 +1,6 @@
 # Exits
 if (!requireNamespace("glmnet", quietly = TRUE)) {
-  exit_file("Package glmnet missing")
+  exit_file("Package 'glmnet' missing")
 }
 
 # # Load required packages
@@ -76,22 +76,4 @@ expect_identical(
 expect_identical(
   current = vip:::get_feature_names.multnet(fit3),
   target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit1, method = "model", include_type = TRUE,
-         mapping = ggplot2::aes(fill = Sign))
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis1, include_type = TRUE),
-  vip(vis2, include_type = TRUE),
-  p,
-  nrow = 1
 )