+++ inst/tinytest/test_pkg_neuralnet.R
@@ -1,9 +1,9 @@
 # Exits
 if (!requireNamespace("neuralnet", quietly = TRUE)) {
-  exit_file("Package neuralnet missing")
+  exit_file("Package 'neuralnet' missing")
 }
 if (!requireNamespace("NeuralNetTools", quietly = TRUE)) {
-  exit_file("Package NeuralNetTools missing")
+  exit_file("Package 'NeuralNetTools' missing")
 }
 
 # # Load required packages
@@ -43,21 +43,4 @@ expect_identical(
 expect_identical(
   current = vip:::get_feature_names.nn(fit),
   target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit, method = "model", include_type = TRUE)
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis1, include_type = TRUE),
-  vip(vis2, include_type = TRUE),
-  p,
-  nrow = 1
 )