+++ inst/tinytest/test_pkg_nnet.R
@@ -1,9 +1,9 @@
 # Exits
 if (!requireNamespace("nnet", quietly = TRUE)) {
-  exit_file("Package nnet missing")
+  exit_file("Package 'nnet' missing")
 }
 if (!requireNamespace("NeuralNetTools", quietly = TRUE)) {
-  exit_file("Package NeuralNetTools missing")
+  exit_file("Package 'NeuralNetTools' missing")
 }
 
 # # Load required packages
@@ -32,33 +32,4 @@ expect_identical(
 expect_identical(
   current = vis2$Importance,
   target = NeuralNetTools::garson(fit, bar_plot = FALSE)$rel_imp
-)
-
-# Expectations for `get_training_data()`
-expect_identical(
-  current = vip:::get_training_data.default(fit),
-  target = friedman1
-)
-
-# Expectations for `get_feature_names()`
-expect_identical(
-  current = vip:::get_feature_names.nnet(fit),
-  target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit, method = "model", include_type = TRUE)
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis1, include_type = TRUE),
-  vip(vis2, include_type = TRUE),
-  p,
-  nrow = 1
 )