+++ inst/tinytest/test_pkg_pls.R
@@ -1,9 +1,9 @@
 # Exits
 if (!requireNamespace("caret", quietly = TRUE)) {
-  exit_file("Package caret missing")
+  exit_file("Package 'caret' missing")
 }
 if (!requireNamespace("pls", quietly = TRUE)) {
-  exit_file("Package pls missing")
+  exit_file("Package 'pls' missing")
 }
 
 # # Load required packages
@@ -37,20 +37,4 @@ expect_identical(
 expect_identical(
   current = vip:::get_feature_names.mvr(fit),
   target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit, method = "model", include_type = TRUE)
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis, include_type = TRUE),
-  p,
-  nrow = 1
 )