+++ inst/tinytest/test_pkg_randomForest.R
@@ -1,6 +1,6 @@
 # Exits
 if (!requireNamespace("randomForest", quietly = TRUE)) {
-  exit_file("Package randomForest missing")
+  exit_file("Package 'randomForest' missing")
 }
 
 # # Load required packages
@@ -66,22 +66,4 @@ expect_identical(
 expect_identical(
   current = vip:::get_feature_names.randomForest(fit3),
   target = paste0("x", 1L:10L)
-)
-
-# Call `vip::vip()` directly
-p <- vip(fit4, method = "model", type = 1, include_type = TRUE)
-
-# Expect `p` to be a `"gg" "ggplot"` object
-expect_identical(
-  current = class(p),
-  target = c("gg", "ggplot")
-)
-
-# Display VIPs side by side
-grid.arrange(
-  vip(vis1, include_type = TRUE),
-  vip(vis2, include_type = TRUE),
-  vip(vis4, include_type = TRUE),
-  p,
-  nrow = 2
 )